% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5.0.R
\name{dgp_y}
\alias{dgp_y}
\title{Generate responses Y from the DGP}
\usage{
dgp_y(x, sigma, type = c("1d", "2d"), seed = NULL)
}
\arguments{
\item{x}{Numeric matrix of covariates.}

\item{sigma}{Noise standard deviation.}

\item{type}{\code{"1d"} or \code{"2d"}.}

\item{seed}{Optional integer.}
}
\value{
Numeric vector.
}
\description{
Generates \eqn{Y = f(X) + \varepsilon},
where \eqn{\varepsilon \sim \mathcal{N}(0,\sigma^2)}.
}
