% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5.0.R
\name{knn_predict}
\alias{knn_predict}
\title{k-NN regression prediction}
\usage{
knn_predict(x_train, y_train, x_test, k)
}
\arguments{
\item{x_train}{Numeric matrix (n x d).}

\item{y_train}{Numeric vector (length n).}

\item{x_test}{Numeric matrix (m x d) or vector (1d).}

\item{k}{Integer.}
}
\value{
Numeric vector of length m.
}
\description{
Computes k-nearest neighbors regression predictions.
}
