% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5.0.R
\name{make_grid}
\alias{make_grid}
\title{Create evaluation grid on \eqn{[0,1]^d}}
\usage{
make_grid(grid_size = 101, d = 1)
}
\arguments{
\item{grid_size}{Integer.}

\item{d}{Dimension (1 or 2).}
}
\value{
List with x_grid and axis.
}
\description{
Create evaluation grid on \eqn{[0,1]^d}
}
