% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5.0.R
\name{plot_mse_1d}
\alias{plot_mse_1d}
\title{Plot MSE decomposition (1d) using ggplot2}
\usage{
plot_mse_1d(res, show = c("bias2", "variance", "mse_y"))
}
\arguments{
\item{res}{Output of mc_knn_decomp() with d = 1.}

\item{show}{Character vector of components.}
}
\value{
ggplot object.
}
\description{
Plot MSE decomposition (1d) using ggplot2
}
